Fix after review (#97)
