+++ test/src/code_field/code_controller_readonly_test.dart
@@ -5,6 +5,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:highlight/languages/java.dart';
 
 import '../common/create_app.dart';
+import '../common/text_editing_value.dart';
 
 const _editableBeginningEnd = '''
 abc
