+++ test/src/history/code_history_controller_test.dart
@@ -9,6 +9,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import '../common/create_app.dart';
 import '../common/snippets.dart';
+import '../common/text_editing_value.dart';
 import '../common/widget_tester.dart';
 
 void main() {
